Update notify command functionality
An optional parameter has been added with which you can specify the country the weather data of which we want to extract.

Also, a method has been added for simple logging of the command execution process (the log file is ignored, but not the folder).

A functionality has been added that, based on the arguments of the command, creates a message about the weather data and sends it with the corresponding notifier.